refactor(client-profile): move resources into Client folder
