+++ backend/app/Http/Middleware/CheckClient.php
@@ -15,7 +15,7 @@ class CheckClient
     {
        $user = Auth::user();
 
-        // تحقق إذا المستخدم مسجل ومدى توفر معلومة أنه عميل (يمكن تغيير الشرط حسب سير العمل الخاص بك)
+
         if (!$user || !$user->clientProfile) {
             return $this->errorResponse('User is not a client or client profile not found.', 403); // 403 Forbidden لأن المستخدم غير مخول الدخول كعميل
         }
