+++ backend/app/Http/Resources/Client/ClientProfileResource.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Http\Resources;
+namespace App\Http\Resources\Client;
 
 use App\Enums\ProjectStatus;
 use Illuminate\Http\Request;
+use App\Http\Resources\UserResource;
 use Illuminate\Http\Resources\Json\JsonResource;
 use App\Http\Resources\Freelancer\ProjectResource;
